+++ src/libs/angular-bootstrap-switch/src/directives/bsSwitch.js
@@ -1,103 +1,44 @@
 'use strict';
 
 angular.module('frapontillo.bootstrap-switch')
-  .directive('bsSwitch', function ($timeout) {
+  .directive('bsSwitch', function ($parse, $timeout) {
     return {
-      restrict: 'EA',
+      restrict: 'A',
       require: 'ngModel',
-      scope: {
-        switchActive: '@',
-        switchOnText: '@',    // changed name
-        switchOffText: '@',   // changed name
-        switchOnColor: '@',   // changed name
-        switchOffColor: '@',  // changed name
-        switchAnimate: '@',
-        switchSize: '@',
-        switchLabel: '@',
-        switchIcon: '@',      // changed behaviour
-        switchWrapper: '@'    // container class modifier
-      },
-      template: function(tElement) {
-        return ('' + tElement.nodeName).toLowerCase() === 'input' ? undefined : '<input>';
-      },
-      replace: true,
       link: function link(scope, element, attrs, controller) {
+        var isInit = false;
 
         /**
-         * Listen to model changes.
+         * Return the true value for this specific checkbox.
+         * @returns {Object} representing the true view value; if undefined, returns true.
          */
-        var listenToModel = function () {
-          // When the model changes
-          controller.$formatters.push(function (newValue) {
-            if (newValue !== undefined) {
-              $timeout(function () {
-                element.bootstrapSwitch('state', newValue || false, true);
-              });
-            }
-          });
-
-          scope.$watch('switchActive', function (newValue) {
-            var active = newValue === true || newValue === 'true' || !newValue;
-            element.bootstrapSwitch('disabled', !active);
-          });
-
-          scope.$watch('switchOnText', function (newValue) {
-            element.bootstrapSwitch('onText', getValueOrUndefined(newValue));
-          });
-
-          scope.$watch('switchOffText', function (newValue) {
-            element.bootstrapSwitch('offText', getValueOrUndefined(newValue));
-          });
-
-          scope.$watch('switchOnColor', function (newValue) {
-            attrs.dataOn = newValue;
-            element.bootstrapSwitch('onColor', getValueOrUndefined(newValue));
-          });
-
-          scope.$watch('switchOffColor', function (newValue) {
-            attrs.dataOff = newValue;
-            element.bootstrapSwitch('offColor', getValueOrUndefined(newValue));
-          });
-
-          scope.$watch('switchAnimate', function (newValue) {
-            element.bootstrapSwitch('animate', scope.$eval(newValue || 'true'));
-          });
-
-          scope.$watch('switchSize', function (newValue) {
-            element.bootstrapSwitch('size', newValue);
-          });
-
-          scope.$watch('switchLabel', function (newValue) {
-            element.bootstrapSwitch('labelText', newValue ? newValue : '&nbsp;');
-          });
-
-          scope.$watch('switchIcon', function (newValue) {
-            if (newValue) {
-              // build and set the new span
-              var spanClass = '<span class=\'' + newValue + '\'></span>';
-              element.bootstrapSwitch('labelText', spanClass);
-            }
-          });
-
-          scope.$watch('switchWrapper', function(newValue) {
-            // Make sure that newValue is not empty, otherwise default to null
-            if (!newValue) {
-              newValue = null;
-            }
-            element.bootstrapSwitch('wrapperClass', newValue);
-          });
+        var getTrueValue = function() {
+          if (attrs.type === 'radio') {
+            return attrs.value || $parse(attrs.ngValue)(scope) || true;
+          }
+          var trueValue = ($parse(attrs.ngTrueValue)(scope));
+          if (angular.isUndefined(trueValue)) {
+            trueValue = true;
+          }
+          return trueValue;
         };
 
         /**
-         * Listen to view changes.
+         * Get a boolean value from a boolean-like string, evaluating it on the current scope.
+         * @param value The input object
+         * @returns {boolean} A boolean value
          */
-        var listenToView = function () {
-          // When the switch is clicked, set its value into the ngModelController's $viewValue
-          element.on('switchChange.bootstrapSwitch', function (e, data) {
-            scope.$apply(function () {
-              controller.$setViewValue(data);
-            });
-          });
+        var getBooleanFromString = function(value) {
+          return scope.$eval(value) === true;
+        };
+
+        /**
+         * Get a boolean value from a boolean-like string, defaulting to true if undefined.
+         * @param value The input object
+         * @returns {boolean} A boolean value
+         */
+        var getBooleanFromStringDefTrue = function(value) {
+          return (value === true || value === 'true' || !value);
         };
 
         /**
@@ -106,29 +47,208 @@ angular.module('frapontillo.bootstrap-switch')
          * @param value The value to check.
          * @returns the original value if it is truthy, {@link undefined} otherwise.
          */
-        var getValueOrUndefined = function(value) {
+        var getValueOrUndefined = function (value) {
           return (value ? value : undefined);
         };
 
-        // Listen and respond to model changes
-        listenToModel();
+        /**
+         * Get the value of the angular-bound attribute, given its name.
+         * The returned value may or may not equal the attribute value, as it may be transformed by a function.
+         *
+         * @param attrName  The angular-bound attribute name to get the value for
+         * @returns {*}     The attribute value
+         */
+        var getSwitchAttrValue = function(attrName) {
+          var map = {
+            'switchRadioOff': getBooleanFromStringDefTrue,
+            'switchActive': function(value) {
+              return !getBooleanFromStringDefTrue(value);
+            },
+            'switchAnimate': getBooleanFromStringDefTrue,
+            'switchLabel': function(value) {
+              return value ? value : '&nbsp;';
+            },
+            'switchIcon': function(value) {
+              if (value) {
+                return '<span class=\'' + value + '\'></span>';
+              }
+            },
+            'switchWrapper': function(value) {
+              return value || 'wrapper';
+            },
+            'switchInverse': getBooleanFromString,
+            'switchReadonly': getBooleanFromString
+          };
+          var transFn = map[attrName] || getValueOrUndefined;
+          return transFn(attrs[attrName]);
+        };
 
-        // Bootstrap the switch plugin
-        element.bootstrapSwitch();
+        /**
+         * Set a bootstrapSwitch parameter according to the angular-bound attribute.
+         * The parameter will be changed only if the switch has already been initialized
+         * (to avoid creating it before the model is ready).
+         *
+         * @param element   The switch to apply the parameter modification to
+         * @param attr      The name of the switch parameter
+         * @param modelAttr The name of the angular-bound parameter
+         */
+        var setSwitchParamMaybe = function(element, attr, modelAttr) {
+          if (!isInit) {
+            return;
+          }
+          var newValue = getSwitchAttrValue(modelAttr);
+          element.bootstrapSwitch(attr, newValue);
+        };
+
+        var setActive = function() {
+          setSwitchParamMaybe(element, 'disabled', 'switchActive');
+        };
+
+        /**
+         * If the directive has not been initialized yet, do so.
+         */
+        var initMaybe = function() {
+          // if it's the first initialization
+          if (!isInit) {
+            var viewValue = (controller.$modelValue === getTrueValue());
+            isInit = !isInit;
+            // Bootstrap the switch plugin
+            element.bootstrapSwitch({
+              radioAllOff: getSwitchAttrValue('switchRadioOff'),
+              disabled: getSwitchAttrValue('switchActive'),
+              state: viewValue,
+              onText: getSwitchAttrValue('switchOnText'),
+              offText: getSwitchAttrValue('switchOffText'),
+              onColor: getSwitchAttrValue('switchOnColor'),
+              offColor: getSwitchAttrValue('switchOffColor'),
+              animate: getSwitchAttrValue('switchAnimate'),
+              size: getSwitchAttrValue('switchSize'),
+              labelText: attrs.switchLabel ? getSwitchAttrValue('switchLabel') : getSwitchAttrValue('switchIcon'),
+              wrapperClass: getSwitchAttrValue('switchWrapper'),
+              handleWidth: getSwitchAttrValue('switchHandleWidth'),
+              labelWidth: getSwitchAttrValue('switchLabelWidth'),
+              inverse: getSwitchAttrValue('switchInverse'),
+              readonly: getSwitchAttrValue('switchReadonly')
+            });
+            if (attrs.type === 'radio') {
+              controller.$setViewValue(controller.$modelValue);
+            } else {
+              controller.$setViewValue(viewValue);
+            }
+          }
+        };
+
+        /**
+         * Listen to model changes.
+         */
+        var listenToModel = function () {
+
+          attrs.$observe('switchActive', function (newValue) {
+            var active = getBooleanFromStringDefTrue(newValue);
+            // if we are disabling the switch, delay the deactivation so that the toggle can be switched
+            if (!active) {
+              $timeout(function() {
+                setActive(active);
+              });
+            } else {
+              // if we are enabling the switch, set active right away
+              setActive(active);
+            }
+          });
+
+          function modelValue() {
+            return controller.$modelValue;
+          }
+
+          // When the model changes
+          scope.$watch(modelValue, function(newValue) {
+            initMaybe();
+            if (newValue !== undefined) {
+              element.bootstrapSwitch('state', newValue === getTrueValue(), false);
+            } else {
+              element.bootstrapSwitch('toggleIndeterminate', true, false);
+            }
+          }, true);
+
+          // angular attribute to switch property bindings
+          var bindings = {
+            'switchRadioOff': 'radioAllOff',
+            'switchOnText': 'onText',
+            'switchOffText': 'offText',
+            'switchOnColor': 'onColor',
+            'switchOffColor': 'offColor',
+            'switchAnimate': 'animate',
+            'switchSize': 'size',
+            'switchLabel': 'labelText',
+            'switchIcon': 'labelText',
+            'switchWrapper': 'wrapperClass',
+            'switchHandleWidth': 'handleWidth',
+            'switchLabelWidth': 'labelWidth',
+            'switchInverse': 'inverse',
+            'switchReadonly': 'readonly'
+          };
+
+          var observeProp = function(prop, bindings) {
+            return function() {
+              attrs.$observe(prop, function () {
+                setSwitchParamMaybe(element, bindings[prop], prop);
+              });
+            };
+          };
+
+          // for every angular-bound attribute, observe it and trigger the appropriate switch function
+          for (var prop in bindings) {
+            attrs.$observe(prop, observeProp(prop, bindings));
+          }
+        };
+
+        /**
+         * Listen to view changes.
+         */
+        var listenToView = function () {
+          if (attrs.type === 'radio') {
+            // when the switch is clicked
+            element.on('change.bootstrapSwitch', function (e) {
+              // discard not real change events
+              if ((controller.$modelValue === controller.$viewValue) && (e.target.checked !== $(e.target).bootstrapSwitch('state'))) {
+                // $setViewValue --> $viewValue --> $parsers --> $modelValue
+                // if the switch is indeed selected
+                if (e.target.checked) {
+                  // set its value into the view
+                  controller.$setViewValue(getTrueValue());
+                } else if (getTrueValue() === controller.$viewValue) {
+                  // otherwise if it's been deselected, delete the view value
+                  controller.$setViewValue(undefined);
+                }
+              }
+            });
+          } else {
+            // When the checkbox switch is clicked, set its value into the ngModel
+            element.on('switchChange.bootstrapSwitch', function (e) {
+              // $setViewValue --> $viewValue --> $parsers --> $modelValue
+              controller.$setViewValue(e.target.checked);
+            });
+          }
+        };
 
         // Listen and respond to view changes
         listenToView();
 
-        // Delay the setting of the state
-        $timeout(function() {
-          element.bootstrapSwitch('state', controller.$modelValue || false, true);
-        });
+        // Listen and respond to model changes
+        listenToModel();
 
         // On destroy, collect ya garbage
         scope.$on('$destroy', function () {
           element.bootstrapSwitch('destroy');
         });
-
       }
+    };
+  })
+  .directive('bsSwitch', function () {
+    return {
+      restrict: 'E',
+      require: 'ngModel',
+      template: '<input bs-switch>',
+      replace: true
     };
   });