+++ src/libs/angular-bootstrap-toggle-switch/test/angular-toggle-switch.spec.js
@@ -1,12 +1,13 @@
 describe('Toggle Switch', function() {
   var $scope, $compile;
 
-  var baseTemplate = '<toggle-switch model="switchState">\n</toggle-switch>';
-  var onLabelTemplate = '<toggle-switch model="switchState" on-label="CUSTOM-ON">\n</toggle-switch>';
-  var offLabelTemplate = '<toggle-switch model="switchState" off-label="CUSTOM-OFF">\n</toggle-switch>';
-  var knobLabelTemplate = '<toggle-switch model="switchState" knob-label="CUSTOM">\n</toggle-switch>';
-  var htmlLabelsTemplate = '<toggle-switch model="switchState" on-label="<i class=\'icon-ok icon-white\'></i>" off-label="<i class=\'icon-remove\'></i>">\n</toggle-switch>';
-  var disabledTemplate = '<toggle-switch model="switchState" is-disabled="isDisabled">\n</toggle-switch>';
+  var baseTemplate = '<toggle-switch ng-model="switchState">\n</toggle-switch>';
+  var onLabelTemplate = '<toggle-switch ng-model="switchState" on-label="CUSTOM-ON">\n</toggle-switch>';
+  var offLabelTemplate = '<toggle-switch ng-model="switchState" off-label="CUSTOM-OFF">\n</toggle-switch>';
+  var knobLabelTemplate = '<toggle-switch ng-model="switchState" knob-label="CUSTOM">\n</toggle-switch>';
+  var htmlLabelsTemplate = '<toggle-switch ng-model="switchState" on-label="<i class=\'icon-ok icon-white\'></i>" off-label="<i class=\'icon-remove\'></i>">\n</toggle-switch>';
+  var disabledTemplate = '<toggle-switch ng-model="switchState" is-disabled="isDisabled">\n</toggle-switch>';
+  var changeTemplate = '<toggle-switch ng-model="switchState" ng-change="changedState()">\n</toggle-switch>';
 
   // Load up just our module
   beforeEach(module('toggle-switch'));
@@ -56,11 +57,19 @@ describe('Toggle Switch', function() {
         $scope.switchState = true;
       });
     });
+    $scope.changedState = function(){};
+    spyOn($scope, 'changedState');
 
     it('changes model to false when clicked', function() {
       var elm = compileDirective(baseTemplate, $scope);
       elm.triggerHandler('click');
       expect($scope.switchState).toEqual(false);
+    });
+
+    it('change handler is called when clicked', function() {
+      var elm = compileDirective(changeTemplate, $scope);
+      elm.triggerHandler('click');
+      expect($scope.changedState).toHaveBeenCalled();
     });
   });
 
@@ -70,12 +79,20 @@ describe('Toggle Switch', function() {
       $scope.$apply(function() {
         $scope.switchState = false;
       });
+      $scope.changedState = function(){};
+      spyOn($scope, 'changedState');
     });
 
     it('changes model to true when clicked', function() {
       var elm = compileDirective(baseTemplate, $scope);
       elm.triggerHandler('click');
       expect($scope.switchState).toEqual(true);
+    });
+
+    it('change handler is called when clicked', function() {
+      var elm = compileDirective(changeTemplate, $scope);
+      elm.triggerHandler('click');
+      expect($scope.changedState).toHaveBeenCalled();
     });
   });